gaussian, improved aperture_grid, created gaussian fit, started sky error
I added a function to represent Gaussian density.
I changed the aperture_grid function to optionally write the ordered
pairs to the file, but always return the list of the ordered pairs
I created a fuction to fit a Gaussian to the negative side of a
histogram of data, which is useful for sky error calculations
I also started the sky_error function, but I need to make my own
photometry function, so I’ll finish that later.